[dqt] Remove old dataquery (#7561)
The old dataquery module has been replaced with the newest dqt module.
