--- modules/dataquery/.gitignore
@@ -1,6 +1,0 @@
-js/react.app.js
-js/react.filterBuilder.js
-js/react.sidebar.js
-js/react.fieldselector.js
-js/react.paginator.js
-js/react.tabs.js
