--- modules/dataquery/php/module.class.inc
@@ -1,60 +1,0 @@
-<?php
-/**
- * This serves as a hint to LORIS that this module is a real module.
- * It does nothing but implement the module class in the module's namespace.
- *
- * PHP Version 5
- *
- * @category   Behavioural
- * @package    Main
- * @subpackage Query
- * @author     Dave MacFarlane <david.macfarlane2@mcgill.ca>
- * @license    http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
- * @link       https://www.github.com/aces/Loris-Trunk/
- */
-namespace LORIS\dataquery;
-
-/**
- * Class module implements the basic LORIS module functionality
- *
- * @category   Behavioural
- * @package    Main
- * @subpackage Query
- * @author     Dave MacFarlane <david.macfarlane2@mcgill.ca>
- * @license    http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
- * @link       https://www.github.com/aces/Loris-Trunk/
- */
-class Module extends \Module
-{
-    /**
-     * {@inheritDoc}
-     *
-     * @param \User $user The user whose access is being checked.
-     *
-     * @return bool whether access is granted
-     */
-    public function hasAccess(\User $user) : bool
-    {
-        return parent::hasAccess($user) && $user->hasPermission('dataquery_view');
-    }
-
-    /**
-     * {@inheritDoc}
-     *
-     * @return string The menu category for this module
-     */
-    public function getMenuCategory() : string
-    {
-        return "Reports";
-    }
-
-    /**
-     * {@inheritDoc}
-     *
-     * @return string The human readable name for this module
-     */
-    public function getLongName() : string
-    {
-        return "Data Query Tool";
-    }
-}
